// backend/src/modules/auth/auth.schema.ts
import { z } from 'zod'

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string().min(1),
})

export const registerSchema = z.object({
  name: z.string().min(2),
  email: z.string().email(),
  password: z.string().min(8),
})

export const storeLoginSchema = z.object({
  storeSlug: z.string().min(1),
  email: z.string().email(),
  password: z.string().min(1),
})